@@ -3,9 +3,15 @@
   "version": "0.7.0",
   "private": true,
   "dependencies": {
-    "@apollo/client": "^3.3.11",
+    "@apollo/client": "^3.7.2",
     "@apollo/react-ssr": "^4.0.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@fast-csv/format": "^4.3.5",
+    "@loadable/babel-plugin": "^5.16.1",
+    "@loadable/component": "^5.15.3",
+    "@loadable/server": "^5.16.1",
+    "@loadable/webpack-plugin": "^5.15.2",
     "@nivo/core": "^0.66.0",
     "@nivo/line": "^0.66.0",
     "@rebass/forms": "^4.0.6",
@@ -30,21 +36,22 @@
     "fs": "^0.0.1-security",
     "googleapis": "^105.0.0",
     "graphql": "^15.5.0",
+    "graphql-scalars": "^1.22.4",
     "history": "^5.0.0",
     "jsonwebtoken": "^8.5.1",
     "moment": "^2.29.1",
     "razzle": "^4.2.17",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-circular-progressbar": "^2.0.4",
     "react-cookie": "^4.1.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.17.1",
+    "react-hook-form": "^7.47.0",
     "react-intl": "^6.5.5",
     "react-loadable": "^5.5.0",
     "react-number-format": "^4.9.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-router-dom": "^6.4.5",
+    "react-scripts": "^5.0.1",
     "react-spring": "^9.2.4",
     "rebass": "^4.0.7",
     "session-file-store": "^1.5.0",
@@ -53,12 +60,12 @@
     "unfetch": "^4.2.0",
     "unique-names-generator": "^4.7.1",
     "universal-cookie-express": "^4.0.3",
-    "yup": "^0.32.9"
+    "yup": "^1.4.0"
   },
   "scripts": {
     "start:tsc": "tsc -b -w --preserveWatchOutput",
-    "start": "concurrently \"yarn start:tsc\" \"razzle start\"",
-    "build": "export NODE_OPTIONS=--max_old_space_size=4096 && tsc -b && razzle build --noninteractive",
+    "start": "NODE_OPTIONS=\"--max_old_space_size=4096 --openssl-legacy-provider\" && concurrently \"yarn start:tsc\" \"razzle start\"",
+    "build": "NODE_OPTIONS=\"--max_old_space_size=4096 --openssl-legacy-provider\" && tsc -b && razzle build --noninteractive",
     "test": "razzle test --env=jsdom",
     "start:prod": "NODE_ENV=production node build/server.js",
     "storybook": "start-storybook -p 6006 -s public",
@@ -90,6 +97,7 @@
   },
   "types": "./razzle.d.ts",
   "devDependencies": {
+    "@babel/core": "^7.22.11",
     "@formatjs/cli": "^6.2.4",
     "@sentry/types": "^6.18.2",
     "@storybook/addon-actions": "^6.5.9",
@@ -108,9 +116,9 @@
     "@types/jest": "^25.2.1",
     "@types/jsonwebtoken": "^8.5.1",
     "@types/node": "^13.13.2",
-    "@types/react": "^17.0.1",
-    "@types/react-dom": "^17.0.1",
-    "@types/react-helmet": "^5.0.11",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-helmet": "^6.1.0",
     "@types/react-router-dom": "^5.1.4",
     "@types/rebass": "^4.0.6",
     "@types/rebass__forms": "^4.0.4",
@@ -126,12 +134,13 @@
     "html-webpack-plugin": "^4.5.2",
     "mini-css-extract-plugin": "^0.9.0",
     "razzle-dev-utils": "^4.2.17",
-    "typescript": "^4.0.3",
+    "typescript": "^4.5.5",
     "webpack": "^4.44.1",
     "webpack-dev-server": "^3.11.2"
   },
   "resolutions": {
-    "@types/react": "^17.0.1",
-    "@types/react-dom": "^17.0.1"
-  }
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9"
+  },
+  "packageManager": "yarn@4.2.2"
 }
