@@ -1,20 +1,35 @@
-const path = require('path');
-const { ReactLoadablePlugin } = require('react-loadable/webpack');
+const LoadableWebpackPlugin = require("@loadable/webpack-plugin");
+const path = require("path");
 
 const appSrc = path.resolve(__dirname, './src');
 
 module.exports = {
+  modifyWebpackConfig(opts) {
+    const config = opts.webpackConfig;
+
+    // add loadable webpack plugin only
+    // when we are building the client bundle
+    if (opts.env.target === "web") {
+      const filename = path.resolve(__dirname, "build");
+
+      // saving stats file to build folder
+      // without this, stats files will go into
+      // build/public folder
+      config.plugins.push(
+        new LoadableWebpackPlugin({
+          outputAsset: false,
+          writeToDisk: { filename },
+        })
+      );
+    }
+    return config;
+  },
   modify: (config, { dev, target }) => {
     let resConfig =
       target === 'web'
         ? {
             ...config,
-            plugins: [
-              ...config.plugins,
-              new ReactLoadablePlugin({
-                filename: './build/react-loadable.json',
-              }),
-            ],
+            plugins: [...config.plugins],
           }
         : {
             ...config,
@@ -56,11 +71,12 @@ module.exports = {
       output: {
         ...resConfig.output,
         devtoolModuleFilenameTemplate: dev
-          ? info =>
+          ? (info) =>
               path
                 .relative(appSrc, info.absoluteResourcePath)
                 .replace(/\\/g, '/')
-          : info => path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
+          : (info) =>
+              path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
       },
       resolve: {
         ...resConfig.resolve,
